+++ starsky/starsky.foundation.platform/Helpers/ArgsHelper.cs
@@ -1044,4 +1044,34 @@ public sealed class ArgsHelper
 
 		return profile;
 	}
+
+	public static List<string> GetRuntimes(List<string> args)
+	{
+		// --runtime
+		var runtimes = new List<string>();
+		const string runtimeDash = "--runtime";
+		const string runtimeDashAndSpace = "--runtime ";
+
+		for ( var arg = 0; arg < args.Count; arg++ )
+		{
+			if ( !args[arg].StartsWith(runtimeDash,
+				     StringComparison.CurrentCultureIgnoreCase) ||
+			     arg + 1 == args.Count )
+			{
+				if ( args[arg].Contains(runtimeDashAndSpace) ) // and space
+				{
+					var runtimeReplaced =
+						args[arg].Replace(runtimeDashAndSpace, string.Empty).Split(',');
+					runtimes.AddRange(runtimeReplaced);
+				}
+
+				continue;
+			}
+
+			var runtime = args[arg + 1].Split(',');
+			runtimes.AddRange(runtime);
+		}
+
+		return runtimes;
+	}
 }