+++ starsky/starsky.foundation.video/GetDependencies/FfMpegDownload.cs
@@ -38,7 +38,19 @@ public class FfMpegDownload : IFfMpegDownload
 		_preflightRunCheck = preflightRunCheck;
 	}
 
-	public async Task<FfmpegDownloadStatus> DownloadFfMpeg()
+	public async Task<List<FfmpegDownloadStatus>> DownloadFfMpeg(List<string> architectures)
+	{
+		var result = new List<FfmpegDownloadStatus>();
+		foreach ( var architecture in
+		         DotnetRuntimeNames.GetArchitecturesNoGenericAndFallback(architectures) )
+		{
+			result.Add(await DownloadFfMpeg(architecture));
+		}
+
+		return result;
+	}
+
+	public async Task<FfmpegDownloadStatus> DownloadFfMpeg(string? architecture = null)
 	{
 		if ( _appSettings.FfmpegSkipDownloadOnStartup == true
 		     || _appSettings is { AddSwaggerExport: true, AddSwaggerExportExitAfter: true } )
@@ -50,10 +62,10 @@ public class FfMpegDownload : IFfMpegDownload
 			return FfmpegDownloadStatus.SettingsDisabled;
 		}
 
-		var currentArchitecture = CurrentArchitecture.GetCurrentRuntimeIdentifier();
-		CreateDirectoryDependenciesFolderIfNotExists(currentArchitecture);
+		architecture ??= CurrentArchitecture.GetCurrentRuntimeIdentifier();
+		CreateDirectoryDependenciesFolderIfNotExists(architecture);
 
-		if ( _hostFileSystemStorage.ExistFile(_ffmpegExePath.GetExePath(currentArchitecture)) )
+		if ( _hostFileSystemStorage.ExistFile(_ffmpegExePath.GetExePath(architecture)) )
 		{
 			return FfmpegDownloadStatus.Ok;
 		}
@@ -66,10 +78,10 @@ public class FfMpegDownload : IFfMpegDownload
 		}
 
 		var binaryIndexBaseUrls = GetCurrentArchitectureIndexUrls(container,
-			currentArchitecture);
+			architecture);
 
 		var downloadStatus =
-			await _downloadBinaries.Download(binaryIndexBaseUrls, currentArchitecture);
+			await _downloadBinaries.Download(binaryIndexBaseUrls, architecture);
 		if ( downloadStatus != FfmpegDownloadStatus.Ok &&
 		     downloadStatus != FfmpegDownloadStatus.OkAlreadyExists )
 		{
@@ -77,12 +89,12 @@ public class FfMpegDownload : IFfMpegDownload
 			return downloadStatus;
 		}
 
-		if ( !await _prepareBeforeRunning.PrepareBeforeRunning(currentArchitecture) )
+		if ( !await _prepareBeforeRunning.PrepareBeforeRunning(architecture) )
 		{
 			return FfmpegDownloadStatus.PrepareBeforeRunningFailed;
 		}
 
-		if ( !await _preflightRunCheck.TryRun(currentArchitecture) )
+		if ( !await _preflightRunCheck.TryRun(architecture) )
 		{
 			return FfmpegDownloadStatus.PreflightRunCheckFailed;
 		}
@@ -103,6 +115,7 @@ public class FfMpegDownload : IFfMpegDownload
 		return path;
 	}
 
+
 	private static KeyValuePair<BinaryIndex?, List<Uri>> GetCurrentArchitectureIndexUrls(
 		FfmpegBinariesContainer container,
 		string currentArchitecture)
@@ -112,7 +125,6 @@ public class FfMpegDownload : IFfMpegDownload
 
 		return new KeyValuePair<BinaryIndex?, List<Uri>>(sortedData, container.BaseUrls);
 	}
-
 
 	private void CreateDirectoryDependenciesFolderIfNotExists(string currentArchitecture)
 	{